Added routines for calibrating LQ parameter
Defines a general class that contains the different functions used during the calibration process, in a similar manner to AoE_cal.py. A few plotting functions are redundant with those found in AoE_cal.py. We should consider generalizing some of these plotting functions in the future for other PSD parameters.